Користувач має можливість переглядати всі створені активні посилання та статистику по них Користувач має можливість переглядати всі створені посилання та статистику по них Користувач може створити нове / видалити існуюче посилання

+++ src/main/java/com/example/demo/service/UserService.java
@@ -68,7 +68,7 @@ public class UserService implements UserDetailsService {
         return userMapper.toDto(user);
     }
 
-    public UserDto updateUser(Long id, UserDto userDto) {
+    public UserDto updateUser(Integer id, UserDto userDto) {
         User user = userRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("User not found"));
